Add block configuration and README
Since Drupal 9.4, Bartik is no longer the default Drupal theme.
Therefore, it's now necessary to include the block configuration in
the config/install directory. (It seems that if you don't otherwise
specify block configuration for a custom theme, installing the theme
means you get a block configuration created that is a copy of that of
the default theme. As long as Bartik was the default theme, we got
what we wanted.)

Add a README noting this, and showing how to create the block
configuration files.